+++ packages/cli/src/commands/create/app/app.ts
@@ -103,7 +103,7 @@ export async function app(env: Env<{name: string}>) {
   });
   workspace.install('react-router-dom', {version: '^5.2.0'});
   workspace.install('@shopify/hydrogen');
-  workspace.install('vite', {dev: true, version: '^2.6.14'});
+  workspace.install('vite', {dev: true, version: '^2.7.0'});
   workspace.install('@vitejs/plugin-react-refresh', {
     dev: true,
     version: '^1.3.2',