+++ packages/hydrogen-plugin-sanity/src/index.ts
@@ -4,7 +4,7 @@ import {
   isClient,
   fetchBuilder,
   graphqlRequestBody,
-  decodeShopifyId
+  decodeShopifyId,
 } from '@shopify/hydrogen';
 export interface UseSanityQueryResponse<T> {
   /** The data returned by the query. */
@@ -25,22 +25,22 @@ export function useSanityQuery<T>({
   variables?: Record<string, any>;
 }): UseSanityQueryResponse<T> {
   if (isClient()) {
-    throw new Error(
-      'Sanity requests should only be made from the server.'
-    );
+    throw new Error('Sanity requests should only be made from the server.');
   }
 
   // @ts-ignore
-  const sanityId = import.meta.env.VITE_SANITY_ID; 
+  const sanityId = import.meta.env.VITE_SANITY_ID;
   // @ts-ignore
   const sanityToken = import.meta.env.VITE_SANITY_TOKEN;
 
   if (!sanityId || !sanityToken) {
-    throw new Error('VITE_SANITY_ID and VITE_SANITY_TOKEN environment variables must exist to make use of useSanityQuery hook.')
+    throw new Error(
+      'VITE_SANITY_ID and VITE_SANITY_TOKEN environment variables must exist to make use of useSanityQuery hook.'
+    );
   }
 
   if (variables.id) {
-    variables.id = decodeShopifyId(variables.id)
+    variables.id = decodeShopifyId(variables.id);
   }
 
   const body = graphqlRequestBody(query, variables);
@@ -57,9 +57,8 @@ export function useSanityQuery<T>({
 
   const {data} = useQuery<UseSanityQueryResponse<T>>(
     [sanityId, body],
-    fetchBuilder<UseSanityQueryResponse<T>>(request),
+    fetchBuilder<UseSanityQueryResponse<T>>(request)
   );
-  
+
   return data as UseSanityQueryResponse<T>;
 }
-