+++ packages/hydrogen/src/components/CartLineQuantityAdjustButton/examples/cart-line-quantity-adjust-button.example.tsx
@@ -1,4 +1,8 @@
-import {CartLineProvider, useCart, CartLineQuantityAdjustButton} from '@shopify/hydrogen';
+import {
+  CartLineProvider,
+  useCart,
+  CartLineQuantityAdjustButton,
+} from '@shopify/hydrogen';
 
 export function App() {
   const {lines} = useCart();
@@ -6,10 +10,16 @@ export function App() {
   return lines.map((line) => {
     return (
       <CartLineProvider key={line.id} line={line}>
-        <CartLineQuantityAdjustButton adjust="increase">Increase quantity</CartLineQuantityAdjustButton>
-        <CartLineQuantityAdjustButton adjust="decrease">Decrease quantity</CartLineQuantityAdjustButton>
-        <CartLineQuantityAdjustButton adjust="remove">Remove from cart</CartLineQuantityAdjustButton>
+        <CartLineQuantityAdjustButton adjust="increase">
+          Increase quantity
+        </CartLineQuantityAdjustButton>
+        <CartLineQuantityAdjustButton adjust="decrease">
+          Decrease quantity
+        </CartLineQuantityAdjustButton>
+        <CartLineQuantityAdjustButton adjust="remove">
+          Remove from cart
+        </CartLineQuantityAdjustButton>
       </CartLineProvider>
-    )
+    );
   });
 }