+++ packages/hydrogen/src/foundation/useShop/use-shop.tsx
@@ -3,7 +3,7 @@ import {ShopifyProviderValue} from '../ShopifyProvider/types';
 let config: ShopifyProviderValue | null = null;
 
 /**
- * The `useShop` hook provides access to values within `shopify.config.js`.The `useShop` hook provides access to values within `shopify.config.js`. It must be a descendent of a `ShopifyProvider` component.
+ * The `useShop` hook provides access to values within `shopify.config.js`. It must be a descendent of a `ShopifyProvider` component.
  */
 export function useShop(): ShopifyProviderValue {
   return config as ShopifyProviderValue;