+++ packages/hydrogen/src/framework/plugins/vite-plugin-hydrogen-middleware.ts
@@ -1,4 +1,4 @@
-import type {Plugin} from 'vite';
+import {Plugin, loadEnv, ResolvedConfig} from 'vite';
 import path from 'path';
 import {promises as fs} from 'fs';
 import {hydrogenMiddleware, graphiqlMiddleware} from '../middleware';
@@ -18,12 +18,14 @@ export default (
      * loading them in an SSR context, rendering them using the `entry-server` endpoint in the
      * user's project, and injecting the static HTML into the template.
      */
-    configureServer(server) {
+    async configureServer(server) {
       const resolve = (p: string) => path.resolve(server.config.root, p);
       async function getIndexTemplate(url: string) {
         const indexHtml = await fs.readFile(resolve('index.html'), 'utf-8');
         return await server.transformIndexHtml(url, indexHtml);
       }
+
+      await polyfillOxygenEnv(server.config);
 
       // The default vite middleware rewrites the URL `/graphqil` to `/index.html`
       // By running this middleware first, we avoid that.
@@ -51,3 +53,24 @@ export default (
     },
   } as Plugin;
 };
+
+declare global {
+  // eslint-disable-next-line no-var
+  var Oxygen: {env: Record<string, string | undefined>; [key: string]: any};
+}
+
+async function polyfillOxygenEnv(config: ResolvedConfig) {
+  const env = await loadEnv(config.mode, config.root, '');
+
+  const publicPrefixes = Array.isArray(config.envPrefix)
+    ? config.envPrefix
+    : [config.envPrefix || ''];
+
+  for (const key of Object.keys(env)) {
+    if (publicPrefixes.some((prefix) => key.startsWith(prefix))) {
+      delete env[key];
+    }
+  }
+
+  globalThis.Oxygen = {env};
+}