+++ packages/hydrogen/src/utilities/index.ts
@@ -20,3 +20,4 @@ export {isServer} from './isServer';
 export {getMeasurementAsParts, getMeasurementAsString} from './measurement';
 export {parseMetafieldValue} from './parseMetafieldValue';
 export {fetchBuilder, graphqlRequestBody, decodeShopifyId} from './fetch';
+export {getTime} from './timing';