+++ packages/playground/server-components-worker/worker.js
@@ -4,6 +4,13 @@ import entrypoint from './src/entry-server.jsx';
 import indexHtml from './dist/client/index.html?raw';
 import {getAssetFromKV} from '@cloudflare/kv-asset-handler';
 
+// Mock Oxygen global
+globalThis.Oxygen = {
+  env: {
+    PRIVATE_VARIABLE: globalThis.PRIVATE_VARIABLE,
+  },
+};
+
 async function assetHandler(event, url) {
   const response = await getAssetFromKV(event, {});